Imported Upstream version 2.4.0

+++ tests/stub_sftp.py
@@ -22,8 +22,11 @@ A stub SFTP server for loopback SFTP testing.
 
 import os
 import sys
-from paramiko import ServerInterface, SFTPServerInterface, SFTPServer, SFTPAttributes, \
-    SFTPHandle, SFTP_OK, AUTH_SUCCESSFUL, OPEN_SUCCEEDED
+
+from paramiko import (
+    ServerInterface, SFTPServerInterface, SFTPServer, SFTPAttributes,
+    SFTPHandle, SFTP_OK, SFTP_FAILURE, AUTH_SUCCESSFUL, OPEN_SUCCEEDED,
+)
 from paramiko.common import o666
 
 
@@ -55,7 +58,7 @@ class StubSFTPHandle (SFTPHandle):
 
 class StubSFTPServer (SFTPServerInterface):
     # assume current folder is a fine root
-    # (the tests always create and eventualy delete a subfolder, so there shouldn't be any mess)
+    # (the tests always create and eventually delete a subfolder, so there shouldn't be any mess)
     ROOT = os.getcwd()
         
     def _realpath(self, path):
@@ -139,11 +142,23 @@ class StubSFTPServer (SFTPServerInterface):
     def rename(self, oldpath, newpath):
         oldpath = self._realpath(oldpath)
         newpath = self._realpath(newpath)
+        if os.path.exists(newpath):
+            return SFTP_FAILURE
         try:
             os.rename(oldpath, newpath)
         except OSError as e:
             return SFTPServer.convert_errno(e.errno)
         return SFTP_OK
+
+    def posix_rename(self, oldpath, newpath):
+        oldpath = self._realpath(oldpath)
+        newpath = self._realpath(newpath)
+        try:
+            os.rename(oldpath, newpath)
+        except OSError as e:
+            return SFTPServer.convert_errno(e.errno)
+        return SFTP_OK
+
 
     def mkdir(self, path, attr):
         path = self._realpath(path)
